feat: add an "Appearance" tab in app settings
closes #104
